+++ src/test/php/net/daringfireball/markdown/unittest/EntityTest.class.php
@@ -1,8 +1,7 @@
 <?php namespace net\daringfireball\markdown\unittest;
 
 use net\daringfireball\markdown\Entity;
-use test\Assert;
-use test\Test;
+use test\{Assert, Test, Values};
 
 class EntityTest extends MarkdownTest {
 
@@ -17,10 +16,30 @@ class EntityTest extends MarkdownTest {
   }
 
   #[Test]
+  public function uppercase_entity() {
+    $this->assertTransformed('<p>&Aacute;</p>', '&Aacute;');
+  }
+
+  #[Test]
+  public function numeric_entity() {
+    $this->assertTransformed('<p>&#60;</p>', '&#60;');
+  }
+
+  #[Test]
+  public function hex_entity() {
+    $this->assertTransformed('<p>It&#x39;s a wrap</p>', 'It&#x39;s a wrap');
+  }
+
+  #[Test]
   public function string_representation() {
     Assert::equals(
       'net.daringfireball.markdown.Entity<&amp;>',
       (new Entity('&amp;'))->toString()
     );
+  }
+
+  #[Test, Values(['&', '1&2', 'Hello & World', '&amp is not a entity;'])]
+  public function unescaped_ampersands_not_confused_with_entity($input) {
+    $this->assertTransformed('<p>'.htmlspecialchars($input).'</p>', $input);
   }
 }